refactoring : token service
